+++ app.rb
@@ -1,5 +1,4 @@
 require_relative 'student'
-require_relative 'teacher'
 require_relative 'book'
 require_relative 'rental'
