+++ main.rb
@@ -1,5 +1,4 @@
 require_relative 'app'
-require_relative 'registerpeople'
 
 class RentBook
   def initialize
@@ -40,6 +39,42 @@ class Main
     print 'Option: '
   end
 
+  def regstudent
+    print 'Age: '
+    age = gets.chomp.to_i
+    print 'Name: '
+    name = gets.chomp
+    print 'Has parent permission? [Y/N]: '
+    parent_permission = gets.chomp.downcase == 'y'
+    print 'Classroom: '
+    classroom = gets.chomp
+    @app.create_student(classroom, age, name, parent_permission)
+  end
+
+  def regteacher
+    print 'Age: '
+    age = gets.chomp.to_i
+    print 'Name: '
+    name = gets.chomp
+    print 'Specialization: '
+    specialization = gets.chomp
+    @app.create_teacher(age, name, specialization)
+  end
+
+  def reg_people
+    print 'Do you want to create a student (1) or a Teacher (2)? [Input the number]:'
+    selection = gets.chomp
+    num = selection.to_i
+    case num
+    when 1
+      regstudent
+    when 2
+      regteacher
+    else
+      puts 'Wrong Input.'
+    end
+  end
+
   def reg_book
     print 'Title: '
     title = gets.chomp
@@ -55,7 +90,7 @@ class Main
   end
 
   def menu
-    until false
+    loop do
       options
       choice = gets.chomp.to_i
       case choice
@@ -64,7 +99,7 @@ class Main
       when 2
         @app.list_people
       when 3
-        Register.new.reg_people
+        reg_people
       when 4
         reg_book
       when 5
